+++ src/app/examples/crud/components/crud.component.ts
@@ -1,6 +1,7 @@
 import { v4 as uuid } from 'uuid';
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { FormBuilder, NgForm } from '@angular/forms';
+import { Router } from '@angular/router';
 import { takeUntil } from 'rxjs/operators';
 import { Subject, Observable } from 'rxjs';
 import { Store, select } from '@ngrx/store';
@@ -9,12 +10,7 @@ import { ROUTE_ANIMATIONS_ELEMENTS } from '@app/core';
 
 import { State } from '../../examples.state';
 import { Book } from '../books.model';
-import {
-  ActionBooksAddOne,
-  ActionBooksDeleteOne,
-  ActionBooksUpdateOne,
-  ActionBooksSelect
-} from '../books.actions';
+import { ActionBooksUpsertOne, ActionBooksDeleteOne } from '../books.actions';
 import { selectSelectedBook, selectAllBooks } from '../books.selectors';
 
 @Component({
@@ -22,7 +18,7 @@ import { selectSelectedBook, selectAllBooks } from '../books.selectors';
   templateUrl: './crud.component.html',
   styleUrls: ['./crud.component.scss']
 })
-export class CrudComponent implements OnInit {
+export class CrudComponent implements OnInit, OnDestroy {
   private unsubscribe$: Subject<void> = new Subject<void>();
   routeAnimationsElements = ROUTE_ANIMATIONS_ELEMENTS;
 
@@ -40,23 +36,35 @@ export class CrudComponent implements OnInit {
     };
   }
 
-  constructor(public store: Store<State>, public fb: FormBuilder) {}
+  constructor(
+    public store: Store<State>,
+    public fb: FormBuilder,
+    private router: Router
+  ) {}
 
   ngOnInit() {
     this.books$ = this.store.pipe(select(selectAllBooks));
     this.store
-      .pipe(select(selectSelectedBook), takeUntil(this.unsubscribe$))
+      .pipe(
+        select(selectSelectedBook),
+        takeUntil(this.unsubscribe$)
+      )
       .subscribe(book => (this.selectedBook = book));
   }
 
-  select(id: string) {
-    this.store.dispatch(new ActionBooksSelect({ id }));
+  ngOnDestroy() {
+    this.unsubscribe$.next();
+    this.unsubscribe$.complete();
+  }
+
+  select(bookId: string) {
     this.isEditing = false;
+    this.router.navigate(['examples/crud', bookId]);
   }
 
   deselect() {
-    this.store.dispatch(new ActionBooksSelect({ id: null }));
     this.isEditing = false;
+    this.router.navigate(['examples/crud']);
   }
 
   edit() {
@@ -68,7 +76,6 @@ export class CrudComponent implements OnInit {
     bookForm.resetForm();
     this.bookFormGroup.reset();
     this.bookFormGroup.setValue(CrudComponent.createBook());
-    this.store.dispatch(new ActionBooksSelect({ id: null }));
     this.isEditing = true;
   }
 
@@ -79,23 +86,15 @@ export class CrudComponent implements OnInit {
   delete() {
     this.store.dispatch(new ActionBooksDeleteOne({ id: this.selectedBook.id }));
     this.isEditing = false;
+    this.router.navigate(['examples/crud']);
   }
 
   save() {
     if (this.bookFormGroup.valid) {
       const book = this.bookFormGroup.value;
-      this.store.dispatch(
-        this.selectedBook
-          ? new ActionBooksUpdateOne({
-              update: {
-                id: book.id,
-                changes: book
-              }
-            })
-          : new ActionBooksAddOne({ book })
-      );
-      this.select(book.id);
+      this.store.dispatch(new ActionBooksUpsertOne({ book }));
       this.isEditing = false;
+      this.router.navigate(['examples/crud', book.id]);
     }
   }
 }