+++ src/app/examples/examples.state.ts
@@ -1,5 +1,4 @@
-import { createFeatureSelector, ActionReducerMap } from '@ngrx/store';
-
+import { ActionReducerMap, createFeatureSelector } from '@ngrx/store';
 import { AppState } from '@app/core';
 
 import { todosReducer } from './todos/todos.reducer';
@@ -7,13 +6,16 @@ import { TodosState } from './todos/todos.model';
 import { stockMarketReducer } from './stock-market/stock-market.reducer';
 import { StockMarketState } from './stock-market/stock-market.model';
 import { bookReducer } from './crud/books.reducer';
-import { BookState } from './crud/books.model';
 import { formReducer } from './form/form.reducer';
-import { FormState } from './form/form.model';import { restaurantReducer } from './restaurants/restaurants.reducer';
+import { FormState } from './form/form.model';
+import { restaurantReducer } from './restaurants/restaurants.reducer';
 import { RestaurantState } from './restaurants/restaurants.model';
+import { Book, BookState } from './crud/books.model';
 
 export const FEATURE_NAME = 'examples';
-
+export const selectExamples = createFeatureSelector<State, ExamplesState>(
+  FEATURE_NAME
+);
 export const reducers: ActionReducerMap<ExamplesState> = {
   todos: todosReducer,
   stocks: stockMarketReducer,
@@ -22,15 +24,11 @@ export const reducers: ActionReducerMap<ExamplesState> = {
   restaurants: restaurantReducer
 };
 
-export const selectExamples = createFeatureSelector<State, ExamplesState>(
-  FEATURE_NAME
-);
-
 export interface ExamplesState {
   todos: TodosState;
   stocks: StockMarketState;
-  books: BookState;
   form: FormState;
+  books: BookState;
   restaurants: RestaurantState;
 }