+++ src/app/examples/form/components/form.component.ts
@@ -1,10 +1,8 @@
-import { Component, OnInit, OnDestroy, ViewChild } from '@angular/core';
+import { Component, OnInit, OnDestroy } from '@angular/core';
 import { Validators, FormBuilder } from '@angular/forms';
-import { CdkTextareaAutosize } from '@angular/cdk/text-field';
-import { MatSnackBar } from '@angular/material/snack-bar';
 import { Store, select } from '@ngrx/store';
 import { Subject } from 'rxjs';
-import { takeUntil, filter, debounceTime, tap } from 'rxjs/operators';
+import { takeUntil, filter, debounceTime, take } from 'rxjs/operators';
 
 import { ROUTE_ANIMATIONS_ELEMENTS } from '@app/core';
 import { TranslateService } from '@ngx-translate/core';
@@ -13,6 +11,7 @@ import { State } from '../../examples.state';
 import { ActionFormUpdate, ActionFormReset } from '../form.actions';
 import { selectForm } from '../form.selectors';
 import { Form } from '../form.model';
+import { NotificationService } from '@app/core/notifications/notification.service';
 
 @Component({
   selector: 'anms-form',
@@ -20,9 +19,7 @@ import { Form } from '../form.model';
   styleUrls: ['./form.component.scss']
 })
 export class FormComponent implements OnInit, OnDestroy {
-  private unsubscribe$: Subject<void> = new Subject<void>();
-
-  @ViewChild('autosize') autosize: CdkTextareaAutosize;
+  private unsubscribe$ = new Subject<void>();
 
   routeAnimationsElements = ROUTE_ANIMATIONS_ELEMENTS;
 
@@ -48,19 +45,21 @@ export class FormComponent implements OnInit, OnDestroy {
     private fb: FormBuilder,
     private store: Store<State>,
     private translate: TranslateService,
-    public snackBar: MatSnackBar
+    private notificationService: NotificationService
   ) {}
 
   ngOnInit() {
     this.store
-      .pipe(select(selectForm))
-      .subscribe(form => this.form.patchValue(form))
-      .unsubscribe();
+      .pipe(
+        select(selectForm),
+        take(1)
+      )
+      .subscribe(form => this.form.patchValue(form.form));
 
     this.form.valueChanges
       .pipe(
         debounceTime(500),
-        filter((form: Form) => form['autosave']),
+        filter((form: Form) => form.autosave),
         takeUntil(this.unsubscribe$)
       )
       .subscribe((form: Form) =>
@@ -77,14 +76,12 @@ export class FormComponent implements OnInit, OnDestroy {
   onSubmit() {
     if (this.form.valid) {
       this.save();
-      this.snackBar.open(
-        this.form.value.requestGift
+      this.notificationService.info(
+        (this.form.value.requestGift
           ? this.translate.instant('anms.examples.form.text4')
-          : this.translate.instant('anms.examples.form.text5'),
-        this.translate.instant('anms.examples.form.text6'),
-        {
-          duration: 1000
-        }
+          : this.translate.instant('anms.examples.form.text5')) +
+          ' : ' +
+          this.translate.instant('anms.examples.form.text6')
       );
     }
   }