+++ src/app/settings/components/settings-container.component.ts
@@ -43,7 +43,10 @@ export class SettingsContainerComponent implements OnInit, OnDestroy {
 
   constructor(private store: Store<{}>) {
     store
-      .pipe(select(selectSettings), takeUntil(this.unsubscribe$))
+      .pipe(
+        select(selectSettings),
+        takeUntil(this.unsubscribe$)
+      )
       .subscribe(settings => (this.settings = settings));
   }